backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pydantic[email]==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
redis==5.0.1
httpx==0.26.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.116.0
websockets==12.0
email-validator==2.1.0
slowapi==0.1.9
bleach==6.1.0
typer[all]>=0.12.0
rich>=13.0.0
psycopg2-binary==2.9.9
